+++ lib/widgets/journal/journal_card.dart
@@ -9,7 +9,6 @@ import 'package:lotti/get_it.dart';
 import 'package:lotti/services/nav_service.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/journal/card_image_widget.dart';
-import 'package:lotti/widgets/journal/entry_details/duration_widget.dart';
 import 'package:lotti/widgets/journal/entry_details/habit_summary.dart';
 import 'package:lotti/widgets/journal/entry_details/health_summary.dart';
 import 'package:lotti/widgets/journal/entry_details/measurement_summary.dart';
@@ -150,12 +149,6 @@ class JournalCardTitle extends StatelessWidget {
               ),
             ),
           ),
-          item.maybeMap(
-            task: (_) => const SizedBox.shrink(),
-            orElse: () => DurationViewWidget(
-              item: item,
-            ),
-          ),
         ],
       ),
     );
@@ -249,7 +242,7 @@ class LeadingIcon extends StatelessWidget {
     return Icon(
       iconData,
       size: 32,
-      color: styleConfig().primaryTextColor,
+      color: styleConfig().secondaryTextColor,
     );
   }
 }